add: RegisterUser.js register new user through React Form
Register User component contains a form to insert user data.

Please not it needed a bit of convincing to get the body object to be
valid JSON.  We need to implement a way to wrap all parameters of a form
into a nice tidy JSON for the flask server.  The way it is now works but
could be better.