refactor: Routes, Loaders & API Calls
# Routes:

Please see src/routes.js for the new route system

Added a profile/me route in addition to the existing profile/:userId to
simplify validation etc.

# Loaders:

Please see the loaders in src/routes.js, these are stored in src/hooks/

Loaders get data from the API and then components can call that data
with useLoaderData() to access the object returned by the loader.

# API Changes:
## Amenities:
Now optionally checks if user is logged in for token validation during
AddPlace step. Note this now returns {amenitites: [], user: 'user_id'}
with user_id being undefined if users jwt is invalid.  This means we can
get a list of all the amenities (say for place results) but we can also
use that call to verify if a user is logged in.

## User:
Getting a user now also gets that users place objects. This means only
one call is necessary for the profile page.

## Place:
Getting a place now also gets that places owner object.  This means only
one call is necessary for the place details page.